LayoutTests/inspector/dom-debugger/command-line-api-event-expected.txt

Checks that $event is available and accurate in evaluations when paused in an event listener.


== Running test suite: CommandLineAPI.$event
-- Running test case: CommandLineAPI.$event.World.Main
Triggering event...
Checking $event...
PASS: Should pause in main world.
Resuming...

-- Running test case: CommandLineAPI.$event.World.Internal
Triggering event...
Checking $event...
PASS: Should pause in internal world.
Resuming...
